dj_filter/requirement.txt
Django==2.1.3
django-filter==2.0.0
django-widget-tweaks==1.4.3
pytz==2018.7
